+++ changelog.md
@@ -1,4 +1,6 @@
 -11 Oct 2018
 ```
 - Add Email Verification by [Bestmomo](https://github.com/bestmomo/laravel-email-confirmation)
+- Add AdminLTE 2 (Alpha v2) by [AdminLTE.io](https://adminlte.io/themes/dev/AdminLTE/)
+- Add new SweetAlert by [t4t5](https://sweetalert.js.org/guides/)
 ```